refactor: back to polyrepo structure
