--- src/packages/v3-plugin/index.ts
@@ -1,2 +1,0 @@
-export { parsers } from './parsers';
-export { printers } from './printers';
