+++ tests/astro/single-plugin.test.ts
@@ -1,0 +1,79 @@
+import { thisPlugin, testEach } from '../adaptor';
+import type { Fixture } from '../settings';
+import { baseOptions } from '../settings';
+
+const options = {
+  ...baseOptions,
+  parser: 'astro',
+};
+
+const fixtures: Fixture[] = [
+  {
+    name: 'astro plugin (1) - standalone use',
+    input: `
+---
+function getDate() {
+    return new Date();
+}
+
+const now = getDate();
+---
+
+<div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
+    <span>Now: {now}</span>
+</div>
+`,
+    output: `---
+function getDate() {
+  return new Date();
+}
+
+const now = getDate();
+---
+
+<div
+  class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl"
+>
+  <span>Now: {now}</span>
+</div>
+`,
+    options: {
+      plugins: ['prettier-plugin-astro'],
+    },
+  },
+  {
+    name: 'astro plugin (2) - a combination of a single plugin and a merge plugin also has no effect',
+    input: `
+---
+function getDate() {
+    return new Date();
+}
+
+const now = getDate();
+---
+
+<div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
+    <span>Now: {now}</span>
+</div>
+`,
+    output: `---
+function getDate() {
+  return new Date();
+}
+
+const now = getDate();
+---
+
+<div
+  class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl"
+>
+  <span>Now: {now}</span>
+</div>
+`,
+    options: {
+      plugins: ['prettier-plugin-astro', thisPlugin],
+    },
+  },
+];
+
+testEach(fixtures, options);
