--- tests/v3-test/angular/zero-plugin.test.ts
@@ -1,56 +1,0 @@
-import type { Fixture } from '../../test-settings';
-import { baseOptions } from '../../test-settings';
-
-import { thisPlugin, testEach } from '../adaptor';
-
-const options = {
-  ...baseOptions,
-  parser: 'angular',
-};
-
-const fixtures: Fixture[] = [
-  {
-    name: 'no plugins (1)',
-    input: `
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: [],
-    },
-  },
-  {
-    name: 'no plugins (2) - merge plugin alone has no effect',
-    input: `
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: [thisPlugin],
-    },
-  },
-];
-
-testEach(fixtures, options);
