--- tests/v3-test/vue/multiple-plugin.test.ts
@@ -1,691 +1,0 @@
-import type { Fixture } from '../../test-settings';
-import { baseOptions } from '../../test-settings';
-
-import {
-  thisPlugin,
-  sortImportsPluginOptions,
-  braceStylePluginOptions,
-  classnamesPluginOptions,
-  testEach,
-} from '../adaptor';
-
-const options = {
-  ...baseOptions,
-  parser: 'vue',
-};
-
-const fixtures: Fixture[] = [
-  {
-    name: 'two plugins whose formatting regions are disjoint (1) - sort-imports -> brace-style',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton
-      @click="
-        () =>
-        {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['@trivago/prettier-plugin-sort-imports', 'prettier-plugin-brace-style', thisPlugin],
-      ...sortImportsPluginOptions,
-      ...braceStylePluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (2) - brace-style -> sort-imports',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton
-      @click="
-        () =>
-        {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-brace-style', '@trivago/prettier-plugin-sort-imports', thisPlugin],
-      ...braceStylePluginOptions,
-      ...sortImportsPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (3) - brace-style -> tailwindcss',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="px-1 font-bold">{{ count }}</span>
-    <CounterButton
-      @click="
-        () =>
-        {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-brace-style', 'prettier-plugin-tailwindcss', thisPlugin],
-      ...braceStylePluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (4) - tailwindcss -> brace-style',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="px-1 font-bold">{{ count }}</span>
-    <CounterButton
-      @click="
-        () =>
-        {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-tailwindcss', 'prettier-plugin-brace-style', thisPlugin],
-      ...braceStylePluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (5) - tailwindcss -> sort-imports',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="px-1 font-bold">{{ count }}</span>
-    <CounterButton
-      @click="
-        () => {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-tailwindcss', '@trivago/prettier-plugin-sort-imports', thisPlugin],
-      ...sortImportsPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (6) - sort-imports -> tailwindcss',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-withDefaults(defineProps<{
-  label?: string;
-}>(), {
-  label: 'Counter',
-})
-
-const count = ref(0)
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="font-bold px-1">{{ count }}</span>
-    <CounterButton @click="() => { count += 1 }" />
-  </CounterContainer>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-withDefaults(
-  defineProps<{
-    label?: string;
-  }>(),
-  {
-    label: "Counter",
-  },
-);
-
-const count = ref(0);
-</script>
-
-<template>
-  <CounterContainer>
-    <span class="px-1">{{ label }}</span>
-    <span class="px-1 font-bold">{{ count }}</span>
-    <CounterButton
-      @click="
-        () => {
-          count += 1;
-        }
-      "
-    />
-  </CounterContainer>
-</template>
-`,
-    options: {
-      plugins: ['@trivago/prettier-plugin-sort-imports', 'prettier-plugin-tailwindcss', thisPlugin],
-      ...sortImportsPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (7) - sort-imports -> classnames',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-const count = ref(0);
-
-function incrementHandler() {
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50
-dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['@trivago/prettier-plugin-sort-imports', 'prettier-plugin-classnames', thisPlugin],
-      ...sortImportsPluginOptions,
-      ...classnamesPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (8) - classnames -> sort-imports',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { ref } from "vue";
-
-import { CounterContainer } from "@/layouts";
-
-import { CounterButton } from "./parts";
-
-const count = ref(0);
-
-function incrementHandler() {
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50
-dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-classnames', '@trivago/prettier-plugin-sort-imports', thisPlugin],
-      ...classnamesPluginOptions,
-      ...sortImportsPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (9) - brace-style -> classnames',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-const count = ref(0);
-
-function incrementHandler()
-{
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50
-dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-brace-style', 'prettier-plugin-classnames', thisPlugin],
-      ...braceStylePluginOptions,
-      ...classnamesPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins whose formatting regions are disjoint (10) - classnames -> brace-style',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-const count = ref(0);
-
-function incrementHandler()
-{
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50
-dark:border-neutral-500/30 px-4 py-4 rounded-xl"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-classnames', 'prettier-plugin-brace-style', thisPlugin],
-      ...classnamesPluginOptions,
-      ...braceStylePluginOptions,
-    },
-  },
-  {
-    name: 'two plugins with some overlapping formatting regions (1) - tailwindcss -> classnames',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-const count = ref(0);
-
-function incrementHandler() {
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="rounded-xl border border-zinc-400/30 bg-gray-100/50 px-4 py-4
-dark:border-neutral-500/30 dark:bg-neutral-900/50"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-tailwindcss', 'prettier-plugin-classnames', thisPlugin],
-      ...classnamesPluginOptions,
-    },
-  },
-  {
-    name: 'two plugins with some overlapping formatting regions (2) - classnames -> tailwindcss',
-    input: `
-<script setup lang="ts">
-import { CounterButton } from './parts'
-import { CounterContainer } from '@/layouts'
-import { ref } from 'vue'
-
-const count = ref(0)
-
-function incrementHandler() {
-  count.value += 1
-}
-</script>
-
-<template>
-  <div class="bg-gray-100/50 border border-zinc-400/30 dark:bg-neutral-900/50 dark:border-neutral-500/30 px-4 py-4 rounded-xl">
-    <slot></slot>
-  </div>
-</template>
-`,
-    output: `<script setup lang="ts">
-import { CounterButton } from "./parts";
-import { CounterContainer } from "@/layouts";
-import { ref } from "vue";
-
-const count = ref(0);
-
-function incrementHandler() {
-  count.value += 1;
-}
-</script>
-
-<template>
-  <div
-    class="rounded-xl border border-zinc-400/30 bg-gray-100/50 px-4 py-4 dark:border-neutral-500/30 dark:bg-neutral-900/50"
-  >
-    <slot></slot>
-  </div>
-</template>
-`,
-    options: {
-      plugins: ['prettier-plugin-classnames', 'prettier-plugin-tailwindcss', thisPlugin],
-      ...classnamesPluginOptions,
-    },
-  },
-];
-
-testEach(fixtures, options);
